backend/requirements.txt
fastapi>=0.109.2
pydantic>=2.7.2,<3.0.0
python-dotenv==1.0.1
uvicorn>=0.27.1
SQLAlchemy>=2.0.27
psycopg2-binary>=2.9.0,<3.0.0
asyncpg>=0.29.0  # Required for async PostgreSQL
alembic>=1.13.1
python-jose[cryptography]<3.4.0,>=3.3.0
passlib[bcrypt]<1.8.0,>=1.7.4
python-multipart>=0.0.9
redis>=5.0.1
bcrypt==4.1.2
aiofiles==23.2.1
httpx==0.27.0
mcp-server==0.1.4
email-validator==2.1.1
websockets>=12.0  # For WebSocket testing
PyJWT>=2.8.0  # Required by security module
anthropic>=0.18.0  # Required for model integration
pydantic-settings>=2.2.1
aiosqlite>=0.19.0
pytz>=2024.1  # Required for timezone handling
aiohttp>=3.9.3 